Move `SemanticModelConfig` to dbt/artifacts
